requirements-startup.txt
aiofiles
aiohttp
aria2p
asyncurban
beautifulsoup4==4.8.1
cfscrape
coffeehouse
covid
cowpy
cryptg
csvfaker
DateTime
emoji
gitpython
google-api-python-client==1.7.11
google_images_download==2.8.0
googletrans
gsearch
gtts
hachoir
heroku3
httplib2
humanize
jotquote
justwatch
kwot
lyricsgenius
oauth2client
patool
Pillow
programmingquotes
psycopg2
pybase64
pyfiglet
PyGithub
pylast
PyLyrics
pySmartDL
python-barcode
python-magic
pytuneteller
qrcode
regex
requests
search-engine-parser
selenium
speedtest-cli
spotify_token
sqlalchemy
telegraph
urbandict
wikipedia
youtube-dl




